fix: correct test relaying on jasmine timeout
The tests run slow, because afterEach was async but never called done(). 
because it was afterEach any errors are ignored and so it did not show
up as failure. Instead the test ran very slow until jasmine timed out.

This fix corrects the afterEach and decreases the timeout to 1,000 ms.